+++ src/aiofiles/os.py
@@ -3,59 +3,58 @@
 import os
 
 from . import ospath as path
-from .base import wrap
+from .base import to_coro
 
 __all__ = [
-    "path",
-    "stat",
+    "access",
+    "getcwd",
+    "listdir",
+    "makedirs",
+    "mkdir",
+    "readlink",
+    "remove",
+    "removedirs",
     "rename",
     "renames",
     "replace",
-    "remove",
-    "unlink",
-    "mkdir",
-    "makedirs",
     "rmdir",
-    "removedirs",
-    "symlink",
-    "readlink",
-    "listdir",
+    "path",
     "scandir",
-    "access",
-    "wrap",
-    "getcwd",
+    "stat",
+    "symlink",
+    "unlink",
 ]
 
-access = wrap(os.access)
+access = to_coro(os.access)
 
-getcwd = wrap(os.getcwd)
+getcwd = to_coro(os.getcwd)
 
-listdir = wrap(os.listdir)
+listdir = to_coro(os.listdir)
 
-makedirs = wrap(os.makedirs)
-mkdir = wrap(os.mkdir)
+makedirs = to_coro(os.makedirs)
+mkdir = to_coro(os.mkdir)
 
-readlink = wrap(os.readlink)
-remove = wrap(os.remove)
-removedirs = wrap(os.removedirs)
-rename = wrap(os.rename)
-renames = wrap(os.renames)
-replace = wrap(os.replace)
-rmdir = wrap(os.rmdir)
+readlink = to_coro(os.readlink)
+remove = to_coro(os.remove)
+removedirs = to_coro(os.removedirs)
+rename = to_coro(os.rename)
+renames = to_coro(os.renames)
+replace = to_coro(os.replace)
+rmdir = to_coro(os.rmdir)
 
-scandir = wrap(os.scandir)
-stat = wrap(os.stat)
-symlink = wrap(os.symlink)
+scandir = to_coro(os.scandir)
+stat = to_coro(os.stat)
+symlink = to_coro(os.symlink)
 
-unlink = wrap(os.unlink)
+unlink = to_coro(os.unlink)
 
 
 if hasattr(os, "link"):
     __all__ += ["link"]
-    link = wrap(os.link)
+    link = to_coro(os.link)
 if hasattr(os, "sendfile"):
     __all__ += ["sendfile"]
-    sendfile = wrap(os.sendfile)
+    sendfile = to_coro(os.sendfile)
 if hasattr(os, "statvfs"):
     __all__ += ["statvfs"]
-    statvfs = wrap(os.statvfs)
+    statvfs = to_coro(os.statvfs)